:recycle:refactor: 게시글 데이터 패칭 로직 컴포넌트로 이동
